refactor: 스케줄 관련 ServiceDto에서 의미 불분명한 Boolean 값이 아닌 Enum 값 가지도록 변경
HomeDateServiceDto, BossWorkspaceDateServiceDto에서
isSchedule, isAfterToday, isAllSuccess(Boolean 타입) ➡️  totalScheduleStatus(Enum 타입) 필드로 변경
